+++ ipwhois/utils.py
@@ -340,6 +340,12 @@ def ipv4_is_defined(address):
 
         return True, 'Private-Use Networks', 'RFC 1918'
 
+    # New IANA Reserved
+    # TODO: Someone needs to find the RFC for this
+    elif query_ip in IPv4Network('198.97.38.0/24'):
+
+        return True, 'IANA Reserved', ''
+
     return False, '', ''